# Huge performance increase in filter parsing