--- a/src/Novell.Directory.Ldap.NETStandard/Utilclass/StringExtensions.cs
+++ b/src/Novell.Directory.Ldap.NETStandard/Utilclass/StringExtensions.cs
@@ -1,0 +1,61 @@
+/******************************************************************************
+* The MIT License
+* Copyright (c) 2003 Novell Inc.,  www.novell.com
+*
+* Permission is hereby granted, free of charge, to any person obtaining  a copy
+* of this software and associated documentation files (the Software), to deal
+* in the Software without restriction, including  without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to  permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in
+* all copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED AS IS, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+* SOFTWARE.
+*******************************************************************************/
+
+//
+// Novell.Directory.Ldap.Utilclass.TokenType.cs
+//
+// Author:
+//   Sunil Kumar (Sunilk@novell.com)
+//
+// (C) 2003 Novell, Inc (http://www.novell.com)
+//
+
+using System;
+
+namespace Novell.Directory.Ldap.Utilclass
+{
+    public static class StringExtensions
+    {
+        /// <summary>
+        /// Replaces string.Substring(offset).StartsWith(value) and avoids memory allocations
+        /// </summary>
+        public static bool StartsWithStringAtOffset(this string baseString, string value, int offset)
+        {
+            if (value == null) throw new ArgumentNullException(nameof(value));
+            if (offset < 0) throw new ArgumentOutOfRangeException(nameof(offset));
+
+            if (offset + value.Length > baseString.Length)
+            {
+                return false;
+            }
+
+            for (int i = 0; i < value.Length; i++)
+            {
+                if (baseString[offset + i] != value[i])
+                    return false;
+            }
+
+            return true;
+        }
+    }
+}