--- a/src/Novell.Directory.Ldap.NETStandard/Utilclass/StringExtensions.cs
+++ b/src/Novell.Directory.Ldap.NETStandard/Utilclass/StringExtensions.cs
@@ -9,8 +9,12 @@ namespace Novell.Directory.Ldap.Utilclass
         /// </summary>
         public static bool StartsWithStringAtOffset(this string baseString, string value, int offset)
         {
-            if (value == null) throw new ArgumentNullException(nameof(value));
-            if (offset < 0) throw new ArgumentOutOfRangeException(nameof(offset));
+            if (value == null) 
+                throw new ArgumentNullException(nameof(value));
+            if (offset < 0) 
+                throw new ArgumentOutOfRangeException(nameof(offset));
+            if (offset > baseString.Length) 
+                throw new ArgumentOutOfRangeException(nameof(offset), "offset cannot be larger than length of string");
 
             if (offset + value.Length > baseString.Length)
             {