We are sometimes using large filters to query information about a lot of active directory groups at once (up to 200 groups). When I developed on my windows machine, the performance was alright and I had the response in about 40ms. When I deployed the .NET Core 3.1 Webservice to a linux machine, the duration for one query went **up to one second**. 

After a lot of performance measuring I figured out that the culprit was the huge number of called `_filter.Substring(_offset).StartsWith(...)` while parsing the filter. This code is much less performant on Linux than on Windows. Besides that there are a lot of memory allocations which can be avoided.

So I replaced those lines with an index based method `StartsWithStringAtOffset`, which I implemented as an extension method on string.

I only benchmarked the **constructor of RfcFilter**. The output of BenchmarkDotNet speaks for itself:

### Windows
| Method |       Runtime |        FilterString |          Mean |       Error |     StdDev | Ratio |     Gen 0 |   Gen 1 | Gen 2 |   Allocated |
|------- |-------------- |-------------------- |--------------:|------------:|-----------:|------:|----------:|--------:|------:|------------:|
|    Old |    .NET 4.7.2 | (&((...)p)) [12517] | 17,807.906 us | 105.4075 us | 98.5982 us | 1.000 | 9687.5000 |       - |     - | 44814.69 KB |
|    New |    .NET 4.7.2 | (&((...)p)) [12517] |    120.077 us |   0.3840 us |  0.3592 us | 0.007 |   33.4473 |       - |     - |   154.35 KB |
|        |               |                     |               |             |            |       |           |         |       |             |
|    Old | .NET Core 3.1 | (&((...)p)) [12517] |  4,336.735 us |  66.2438 us | 61.9644 us |  1.00 | 9703.1250 | 23.4375 |     - | 44680.05 KB |
|    New | .NET Core 3.1 | (&((...)p)) [12517] |    111.288 us |   0.4098 us |  0.3834 us |  0.03 |   33.0811 |  5.4932 |     - |   152.32 KB |
|        |               |                     |               |             |            |       |           |         |       |             |
|    Old | .NET Core 5.0 | (&((...)p)) [12517] |  8,095.366 us |  98.9044 us | 92.5152 us |  1.00 | 9703.1250 | 31.2500 |     - | 44680.05 KB |
|    New | .NET Core 5.0 | (&((...)p)) [12517] |    111.475 us |   0.6306 us |  0.5899 us |  0.01 |   33.0811 |  5.4932 |     - |   152.32 KB |
|        |               |                     |               |             |            |       |           |         |       |             |
|    Old |    .NET 4.7.2 |  (&((...)up)) [105] |      7.001 us |   0.0151 us |  0.0141 us |  1.00 |    1.7395 |       - |     - |     8.03 KB |
|    New |    .NET 4.7.2 |  (&((...)up)) [105] |      1.284 us |   0.0040 us |  0.0037 us |  0.18 |    0.4463 |       - |     - |     2.06 KB |
|        |               |                     |               |             |            |       |           |         |       |             |
|    Old | .NET Core 3.1 |  (&((...)up)) [105] |      4.494 us |   0.0176 us |  0.0147 us |  1.00 |    1.7014 |  0.0076 |     - |     7.82 KB |
|    New | .NET Core 3.1 |  (&((...)up)) [105] |      1.187 us |   0.0051 us |  0.0047 us |  0.26 |    0.4406 |  0.0019 |     - |     2.03 KB |
|        |               |                     |               |             |            |       |           |         |       |             |
|    Old | .NET Core 5.0 |  (&((...)up)) [105] |     25.611 us |   0.1260 us |  0.1179 us |  1.00 |    1.6785 |       - |     - |     7.82 KB |
|    New | .NET Core 5.0 |  (&((...)up)) [105] |      1.138 us |   0.0055 us |  0.0051 us |  0.04 |    0.4406 |  0.0019 |     - |     2.03 KB |

Interestingly the performance was even worse on .NET 5 compared to .NET Core 3.1 despite all the performance improvements in the framework.

### Linux (only .NET Core 3.1)

| Method |        FilterString |           Mean |         Error |        StdDev | Ratio |     Gen 0 |  Gen 1 | Gen 2 |   Allocated |
|------- |-------------------- |---------------:|--------------:|--------------:|------:|----------:|-------:|------:|------------:|
|    Old | (&((...)p)) [11853] | 473,058.908 us | 4,711.7555 us | 4,176.8485 us | 1.000 | 7000.0000 |      - |     - | 34547.02 KB |
|    New | (&((...)p)) [11853] |     120.620 us |     1.1843 us |     1.0498 us | 0.000 |   28.8086 | 4.0283 |     - |   132.66 KB |
|        |                     |                |               |               |       |           |        |       |             |
|    Old |  (&((...)up)) [109] |       5.076 us |     0.0330 us |     0.0276 us |  1.00 |    1.7395 | 0.0076 |     - |     8.02 KB |
|    New |  (&((...)up)) [109] |       1.520 us |     0.0030 us |     0.0024 us |  0.30 |    0.4444 | 0.0019 |     - |     2.05 KB |

The BenchmarkDotNet for windows code was

```csharp
[SimpleJob(RuntimeMoniker.Net472)]
[SimpleJob(RuntimeMoniker.NetCoreApp31)]
[SimpleJob(RuntimeMoniker.NetCoreApp50)]
[MemoryDiagnoser]
public class Benchmark
{
    public IEnumerable<string> FilterStrings => new List<string>
    {
        "(&(|(distinguishedName=CN=My Group,OU=Central,OU=Groups,DC=domain,DC=example,DC=com))(objectClass=group))",
        // Search for 150 groups
        $"(&(|{string.Join("", Enumerable.Range(1, 150).Select(n => $"CN=My Group {n},OU=Central,OU=Groups,DC=domain,DC=example,DC=com").Select(g => $"(distinguishedName={g})"))})(objectClass=group))"
    };

    [ParamsSource(nameof(FilterStrings))]
    public string FilterString;

    [Benchmark(Baseline = true)]
    public void Old()
    {
        new RfcFilterCurrent(FilterString);
    }

    [Benchmark]
    public void New()
    {
        new RfcFilter(FilterString);
    }
}
```

The code for linux was pretty similar but I only ran it with `[SimpleJob(RuntimeMoniker.NetCoreApp31)]` and the filter contained other groups, hence the small difference in size.